+++ learning/katas/go/core_transforms/branching/branching/cmd/main.go
@@ -16,7 +16,7 @@
 package main
 
 import (
-	"branching/pkg/task"
+	"beam.apache.org/learning/katas/core_transforms/branching/branching/pkg/task"
 	"context"
 	"github.com/apache/beam/sdks/go/pkg/beam"
 	"github.com/apache/beam/sdks/go/pkg/beam/log"
