+++ learning/katas/go/core_transforms/branching/branching/test/task_test.go
@@ -16,7 +16,7 @@
 package test
 
 import (
-	"branching/pkg/task"
+	"beam.apache.org/learning/katas/core_transforms/branching/branching/pkg/task"
 	"github.com/apache/beam/sdks/go/pkg/beam"
 	"github.com/apache/beam/sdks/go/pkg/beam/testing/passert"
 	"github.com/apache/beam/sdks/go/pkg/beam/testing/ptest"
