+++ learning/katas/go/core_transforms/branching/lesson-info.yaml
@@ -18,4 +18,4 @@
 #
 
 content:
-- Branching
+- branching
