+++ learning/katas/go/core_transforms/branching/lesson-remote-info.yaml
