+++ learning/katas/go/core_transforms/section-info.yaml
@@ -26,4 +26,4 @@ content:
 - partition
 - side_input
 - additional_outputs
-- Branching
+- branching
